% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addChr2Exclude}
\alias{addChr2Exclude}
\title{Add a globally-applied requirement for filtering out scaffolds from
all analysis}
\usage{
addChr2Exclude(chr2exclude = c("chrM", "MT", "Pltd", "chrPltd"))
}
\arguments{
\item{chr2exclude}{A character vector, NA or NULL, specifying chromosomes or
scaffolds to be excluded for InPAS analysis. \code{chrM} and alternative scaffolds
representing different haplotypes should be excluded.}
}
\description{
This function will set the default requirement of filtering out scaffolds
from all analysis.
}
