% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addInPASEnsDb}
\alias{addInPASEnsDb}
\title{Add a globally defined EnsDb to some InPAS functions.}
\usage{
addInPASEnsDb(EnsDb = NULL)
}
\arguments{
\item{EnsDb}{An object of \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}}
}
\description{
Add a globally defined EnsDb to some InPAS functions.
}
