% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addInPASGenome}
\alias{addInPASGenome}
\title{Add a globally defined genome to all InPAS functions.}
\usage{
addInPASGenome(genome = NULL)
}
\arguments{
\item{genome}{A BSgenome object indicating the default genome to be
used for all InPAS functions. This value is stored as a global environment
variable. This can be overwritten on a per-function basis using the given
function's \code{genome} parameter.}
}
\description{
This function will set the genome across all InPAS functions.
}
