% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addInPASOutputDirectory}
\alias{addInPASOutputDirectory}
\title{Add a globally defined output directory to some InPAS functions.}
\usage{
addInPASOutputDirectory(outdir = NULL)
}
\arguments{
\item{outdir}{A character(1) vector, a path with write permission for storing
InPAS analysis results. If it doesn't exist, it will be created.}
}
\description{
Add a globally defined output directory to some InPAS functions.
}
