% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addInPASTxDb}
\alias{addInPASTxDb}
\title{Add a globally defined TxDb for InPAS functions.}
\usage{
addInPASTxDb(TxDb = NULL)
}
\arguments{
\item{TxDb}{An object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}
}
\description{
Add a globally defined TxDb for InPAS functions.
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
addInPASTxDb(TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene)
}
