% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{addLockName}
\alias{addLockName}
\title{Add a filename for locking a SQLite database}
\usage{
addLockName(filename = NULL)
}
\arguments{
\item{filename}{A character(1) vector, specifyong a path to a file for locking.}
}
\description{
Add a filename for locking a SQLite database
}
