% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bam2RleCov.R
\name{Bam2RleCov}
\alias{Bam2RleCov}
\title{Get read coverage in the Rle format from a BAM file}
\usage{
Bam2RleCov(
  bamfile,
  index = bamfile,
  tag,
  param = NULL,
  mapqFilter = 255,
  run.type = c("single_end", "paired_end"),
  removeScaffolds = FALSE,
  hugeData = TRUE,
  outdir = NULL
)
}
\arguments{
\item{bamfile}{A character(1) file name of the 'BAM' file to be processed.}

\item{index}{A character(1) name of the index file of the 'BAM' file being
processed; this is given without the '.bai' extension}

\item{tag}{A character(1) name tags used to label the input bamfile}

\item{param}{An instance of \link[Rsamtools:ScanBamParam-class]{Rsamtools::ScanBamParam}. This influences
what fields and which records are imported. Use of which requires that a
BAM index file (filename.bai) exists.}

\item{mapqFilter}{An integer(1), mapping quality filtering threshold.
Default is 255, good for uniquely mapping reads by RNA aligner, STAR.}

\item{run.type}{A character(1), sequencing read layout: single end or
paired end reads?}

\item{removeScaffolds}{A logical(1), whether the scaffolds should be removed
from the genome}

\item{hugeData}{A logical(1). Is this dataset consume too much memory? if it is
TRUE, the coverage will be saved into tempfiles.}

\item{outdir}{A directory with write permission for storing intermediate
coverage data. This is required for huge data, i.e., when hugeData is set
to TRUE}
}
\value{
If hugeData is set "FALSE", the function returns a list containing
read coverage and total depth for a bedgraph file. The first list element
contains a list of Rle instances of \link[S4Vectors:Rle-class]{S4Vectors::Rle} representing
read coverage for each chromosome, with the chromosome names starting
with "chr" as names. The second element contains total depth. Otherwise,
the returned value is a character(1), with tag as the name.
}
\description{
Get read coverage in the Rle format from a BAM file
}
