% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedgraph2RleCov.R
\name{bedgraph2RleCov}
\alias{bedgraph2RleCov}
\title{Convert coverage of a single sample in bedgraph format to Rle}
\usage{
bedgraph2RleCov(
  bedgraph,
  tag,
  sqlite_db,
  genome,
  removeScaffolds = FALSE,
  hugeData = FALSE,
  outdir = NULL
)
}
\arguments{
\item{bedgraph}{A character(1) of a file name of bedgraph generated by
bedtools. eg: bedtools genomecov -bga -split -ibam $bam > $bedgraph}

\item{tag}{Name tag used to label each input bedgraph}

\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{hugeData}{logical(1). Is this dataset consume too much memory? if it is
TRUE, the coverage will be saved into tempfiles.}

\item{outdir}{A path to a directory/folder where output is written}
}
\value{
If hugeData is set "FALSE", the function returns a list containing
read coverage and total depth for a bedgraph file. The first list element
contains a list of Rle instances of \link[S4Vectors:Rle-class]{S4Vectors::Rle} representing
read coverage for each chromosome, with the chromosome names starting
with "chr" as names. The second element contains total depth. Otherwise,
the returned value is a character(1), with tag as the name.
}
\description{
Convert coverage of a single sample in the bedgraph format to an object of
\link[S4Vectors:Rle-class]{S4Vectors::Rle}
}
