% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.calculate_mse.R
\name{calculate_mse}
\alias{calculate_mse}
\title{Calculate mean squared errors (MSE)}
\usage{
calculate_mse(.ele, search_point_START, search_point_END)
}
\arguments{
\item{.ele}{A numeric vector, storing 3' UTR coverage for a give sample or
collapsed 3' UTR coverage for a given condition}

\item{search_point_START}{An integer, specifying the start position to
calculate MSE}

\item{search_point_END}{An integer, specifying end position to calculate MSE}
}
\value{
a vector of numeric, containing mean squared errors for each searched
site when which is assumed as a bisection site (i.e. potential CP site).
}
\description{
Calculate mean squared errors (MSE) for each searched site which is assumed
bisection site (i.e. potential CP site).
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
