% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3TotalCoverage.R
\name{compensation}
\alias{compensation}
\title{Compensate the coverage with GC-content or mappability}
\usage{
compensation(view, comp, start, end)
}
\arguments{
\item{view}{}

\item{comp}{}

\item{start}{}

\item{end}{}
}
\value{

}
\description{
Compensate the coverage with GC-content or mappability
}
\keyword{internal}
