% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr3UsageEstimation.R
\name{coverageCDS}
\alias{coverageCDS}
\title{calculate the coverage for the last CDSs}
\usage{
coverageCDS(coverage, CDS, genome, gp1, gp2, hugeData)
}
\arguments{
\item{coverage}{coverage for each sample, outputs of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{CDS}{a object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} for last CDSs}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{gp1}{tag names involved in group 1}

\item{gp2}{tag names involved in group 2}

\item{hugeData}{logical(1), indicating whether the data is huge?}
}
\value{
a data frame containing CDS total coverage
}
\description{
calculate the coverage for the last CDSs
}
