% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covThreshold.R
\name{covThreshold}
\alias{covThreshold}
\title{calculate the cutoff threshold of coverage}
\usage{
covThreshold(
  coverage,
  genome,
  TxDb,
  removeScaffolds = FALSE,
  utr3,
  chr = "chr1",
  hugeData,
  groupList
)
}
\arguments{
\item{coverage}{Coverage for each sample, output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{TxDb}{An object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}

\item{utr3}{Output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{chr}{Chromosome to be used for calculation, default is "chr1"}

\item{hugeData}{Is this dataset consume too much memory? if it is TRUE, the
coverage will be saved into tempfiles.}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}
}
\value{
A numeric vector
}
\description{
calculate the cutoff threshold of coverage for long and short isoforms
}
\details{
automatically determining the long_coverage_threshold by non_zero
intergenicRegion coverageauto determine the short_coverage_threshold by
quantile of non_zero intragenicRegion coverage
}
\keyword{internal}
