% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthWeight.R
\name{depthWeight}
\alias{depthWeight}
\title{calculate the depth weight for each example}
\usage{
depthWeight(coverage, hugeData, groupList = NULL)
}
\arguments{
\item{coverage}{a list. output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{hugeData}{is it a huge dataset?}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}
}
\value{
a numeric vector with depth weight
}
\description{
calculate the depth weight for each example: depth/mean(depth)
}
\keyword{internal}
