% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distalAdj.R
\name{distalAdj}
\alias{distalAdj}
\title{adjust distal CP sites by cleanUpdTSeq}
\usage{
distalAdj(
  distalCPs,
  classifier,
  classifier_cutoff,
  shift_range,
  genome,
  step = 1
)
}
\arguments{
\item{distalCPs}{the output of \code{\link[=searchDistalCPs]{searchDistalCPs()}}}

\item{classifier}{cleanUpdTSeq classifier}

\item{classifier_cutoff}{cutoff value of the classifier}

\item{shift_range}{the searching range for the better CP sites}

\item{genome}{a \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object}

\item{step}{adjusting step, default 1, means adjusting by each base by
cleanUpdTSeq.}
}
\description{
adjust distal CP sites by cleanUpdTSeq
}
\seealso{
\code{\link[=searchProximalCPs]{searchProximalCPs()}}, \code{\link[=PAscore2]{PAscore2()}}
}
\keyword{internal}
