% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{.onAttach}
\alias{.onAttach}
\title{A function called upon a package is attached to the search path}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
A function called upon a package is attached to the search path
}
\keyword{internal}
