% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3TotalCoverage.R
\name{fft.smooth}
\alias{fft.smooth}
\title{Smoothing using Fast Discrete Fourier Transform}
\usage{
fft.smooth(sn, p)
}
\arguments{
\item{sn}{a real or complex array containing the values to be transformed.
see \code{\link[stats:fft]{stats::fft()}}}

\item{p}{An integer(1), fft smoothing power}
}
\value{
a numeric vector, the real part of inverse fft-transformed signal
}
\description{
Smoothing using Fast Discrete Fourier Transform
}
\keyword{internal}
