% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.find_minMSEDistr.R
\name{find_minMSEDistr}
\alias{find_minMSEDistr}
\title{Visualization of MSE profiles, 3' UTR coverage and minimal MSE distribution}
\usage{
find_minMSEDistr(
  CPs,
  outdir = NULL,
  MSE.plot = "MSE.pdf",
  coverage.plot = "coverage.pdf",
  min.MSE.to.end.distr.plot = "min.MSE.to.end.distr.pdf"
)
}
\arguments{
\item{CPs}{A list, output from \code{\link[=search_proximalCPs]{search_proximalCPs()}} or \code{\link[=adjust_distalCPs]{adjust_distalCPs()}}
or \code{\link[=adjust_proximalCPs]{adjust_proximalCPs()}}}

\item{outdir}{A character(1) vector, specifying the output directory}

\item{MSE.plot}{A character(1) vector, specifying a PDF file name for
outputting plots of MSE profiles. No directory path is allowed.}

\item{coverage.plot}{A character(1) vector, specifying a PDF file name for
outputting per-sample coverage profiles. No directory path is allowed.}

\item{min.MSE.to.end.distr.plot}{A character(1) vector, specifying a PDF
file name for outputting histograms showing minimal MSE distribution
relative to longer 3' UTR end. No directory path is allowed.}
}
\description{
Visualization of MSE profiles, 3' UTR coverage and minimal MSE distribution
}
