% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.find_segmentationSites.R
\name{find_segmentationSites}
\alias{find_segmentationSites}
\title{Find optimal segmentation sites by minimizing MSE}
\usage{
find_segmentationSites(
  .ele,
  search_point_START,
  search_point_END,
  n = 1,
  savedID = NA
)
}
\arguments{
\item{.ele}{3' UTR coverage}

\item{search_point_START}{start position to calculate}

\item{search_point_END}{end position to calculate}

\item{n}{the length of output. If set to -1, return all candidate CP sites}

\item{savedID}{the proximal CPsites}
}
\value{
a list of MSE and candidate APA sites
}
\description{
Find optimal segmentation sites by minimizing MSE
}
\author{
Jianhong Ou
}
\keyword{internal}
