% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.find_valley.R
\name{find_valley}
\alias{find_valley}
\title{find the local minimal mean squared error (MSE)}
\usage{
find_valley(x, ss, se, n = 1, savedID = NA, filterByPval = TRUE)
}
\arguments{
\item{x}{a numeric(n) vector containing MSEs for a given range.}

\item{ss}{position to start searching positions of local minimal MSE}

\item{se}{position to end searching positions of local minimal MSE}

\item{n}{the length of output. If n=-1, output all the local minimal MSE
positions.}

\item{savedID}{saved positions of local minimal MSE}

\item{filterByPval}{A logical(1) vector, indicating whether tp filter the
candidate positions by significant levels (p values) or not.}
}
\value{
a numeric vector containing a number of candidate CP sites.
}
\description{
for a giving numeric vectors, calculate the top N local minimal mean squared
error. The sites of local minimal MSE are saved if they are in the range of
searching start site to searching end site
}
\author{
Jianhong Ou
}
\keyword{internal}
