% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherExactTest.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{do Fisher Exact Test for two group datasets}
\usage{
fisherExactTest(UTR3eset, gp1, gp2)
}
\arguments{
\item{UTR3eset}{output of \code{\link[=getUTR3eSet]{getUTR3eSet()}}}

\item{gp1}{tag names of group 1}

\item{gp2}{tag names of group 2}
}
\value{
a matrix of test results
}
\description{
do Fisher Exact Test for two group datasets
}
\examples{
load(system.file("extdata", "eset.MAQC.rda", package = "InPAS"))
sample_tags <- colnames(eset$PDUI.log2)
res <- fisherExactTest(eset,
  gp1 = sample_tags[1:2],
  gp2 = sample_tags[3:4]
)
}
\seealso{
\code{\link[=singleSampleAnalysis]{singleSampleAnalysis()}} for a single-sample APA
analysis,\code{\link[=singleGroupAnalysis]{singleGroupAnalysis()}} for a single-group sample APA analysis,
\code{\link[=limmaAnalysis]{limmaAnalysis()}} for limma-based APA analysis of complex experimental
design
}
