% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.get_ssRleCov.R
\name{fix_dbLockError}
\alias{fix_dbLockError}
\title{Helper function for insistent database updating}
\usage{
fix_dbLockError()
}
\value{
A function: insist_execute_sqlite()
}
\description{
Helper function for insistent database updating
}
\references{
https://github.com/r-dbi/RSQLite/issues/280. krlmlr commented
on Dec 27, 2020
}
\author{
Haibo Liu
}
\keyword{internal}
