% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{gcComp}
\alias{gcComp}
\title{Calculate weights for GC composition}
\usage{
gcComp(genome, seqnames, window = 50)
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{seqnames}{chromosome/scaffold names in the same forms of seqnames in
the genome}

\item{window}{size of a sliding window, which optimally is set to the read
length}
}
\value{
A list of numeric vectors containing the weight (scaffold-level GC\\%
/ GC\\% per sliding window) for GC composition-based correction (Cheung et
al. 2011).
}
\description{
Calculate read weights for GC composition-based coverage correction
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
genome <- BSgenome.Mmusculus.UCSC.mm10
InPAS:::gcComp(genome, "chr1")
}
\references{
Cheung MS, Down TA, Latorre I, Ahringer J. Systematic bias in
high-throughput sequencing data and its correction by BEADS. Nucleic Acids
Res. 2011 Aug;39(15):e103. doi: 10.1093/nar/gkr425. Epub 2011 Jun 6. PubMed
PMID: 21646344;
}
\keyword{internal}
