% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{gcContents}
\alias{gcContents}
\title{Title}
\usage{
gcContents(genome, seqname, nonATCGExclude = TRUE)
}
\arguments{
\item{genome}{an object of [BSgenome::BSgenome-class]}

\item{seqname}{a character (1) vector, the chromosome/scaffold's
name}

\item{nonATCGExclude}{a logical(1) vector, whether nucleotides 
other than A, T, C, and G should be excluded when GC content 
is calculated}
}
\value{
a numeric(1) vector, containing the chromosome/scaffold
-specific GC content in the range of [0, 1]
}
\description{
Title
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
genome <- BSgenome.Mmusculus.UCSC.mm10
gcContents(genome, "chr1")

}
