% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3UTRCDSRegionCoverage.R
\name{get3UTRCDSRegionCoverage}
\alias{get3UTRCDSRegionCoverage}
\title{Get coverage for 3' UTR regions of a given chromosome}
\usage{
get3UTRCDSRegionCoverage(
  seqname,
  utr3.cds.regions.chr,
  hugeData = TRUE,
  coverage,
  tmpfolder = NULL,
  phmm = FALSE
)
}
\arguments{
\item{seqname}{chromosome/scaffold name}

\item{utr3.cds.regions.chr}{An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}, with
a GRanges for each seqname (chromosome/scaffold). It is the output from the
\code{\link[=get3UTRCDSRegions]{get3UTRCDSRegions()}}}

\item{hugeData}{logical(1), indicating whether it is huge data}

\item{coverage}{coverage for each sample, output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}},
\code{\link[=getCov4SmallExperiment]{getCov4SmallExperiment()}}, or \code{\link[=getCovFileList]{getCovFileList()}}}

\item{tmpfolder}{A path to a folder for storing coverage data of 3' UTRs on a given
given chromosome/scaffold}

\item{phmm}{logical(1), indicatin whether data should be prepared for singleSample
analysis? By default, FALSE}
}
\value{
coverage view
}
\description{
Get coverage for 3' UTR regions of a given chromosome
}
