% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3UTRCDSRegionCoveragePerSeqname.R
\name{get3UTRCDSRegionCoveragePerSeqname}
\alias{get3UTRCDSRegionCoveragePerSeqname}
\title{Get coverage for 3' UTR and last CDS regions on a single chromosome}
\usage{
get3UTRCDSRegionCoveragePerSeqname(
  utr3.cds.regions.chr,
  hugeData = TRUE,
  coverage,
  tmpfolder = NULL,
  phmm = FALSE
)
}
\arguments{
\item{utr3.cds.regions.chr}{An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList},
with a GRanges for a single seqname (chromosome/scaffold). It is an
output from the \code{\link[=get3UTRCDSRegionsPerSeqname]{get3UTRCDSRegionsPerSeqname()}}}

\item{hugeData}{logical(1), indicating whether it is huge data}

\item{coverage}{coverage for each sample, output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}},
\code{\link[=getCov4SmallExperiment]{getCov4SmallExperiment()}}, or \code{\link[=getCovFileList]{getCovFileList()}}}

\item{tmpfolder}{A path to a folder for storing coverage data of 3' UTRs on a
given chromosome/scaffold}

\item{phmm}{logical(1), indicating whether data should be prepared for
singleSample analysis? By default, FALSE}
}
\value{
coverage view
}
\description{
Get coverage for 3' UTR and last CDS regions on a single chromosome
}
