% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3UTRCDSRegions.R
\name{get3UTRCDSRegions}
\alias{get3UTRCDSRegions}
\title{Get 3' UTR regions based on CP sites}
\usage{
get3UTRCDSRegions(CPsites, utr3, BPPARAM = NULL)
}
\arguments{
\item{CPsites}{CP sites detected by \code{\link[=CPsites]{CPsites()}} or \code{\link[=integrateCPsites]{integrateCPsites()}}}

\item{utr3}{output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply.}
}
\value{
An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}, with a GRanges for
each seqname (chromosome/scaffold)
}
\description{
Get 3' UTR regions based on CP sites
}
