% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3UTRCDSRegionsPerSeqname.R
\name{get3UTRCDSRegionsPerSeqname}
\alias{get3UTRCDSRegionsPerSeqname}
\title{Get 3' UTRs and their last CDS regions based on CP sites}
\usage{
get3UTRCDSRegionsPerSeqname(CPsites, utr3, BPPARAM = NULL)
}
\arguments{
\item{CPsites}{CP sites on a single chromosome/scaffold, as detected by
\code{\link[=getCPsitesPerSeqname]{getCPsitesPerSeqname()}}}

\item{utr3}{output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply.}

\item{seqname}{character(1), chromosome/scaffold name}
}
\value{
An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}, with a GRanges for
a seqname (chromosome/scaffold)
}
\description{
Get 3' UTRs and their last CDS regions based on CP sites
}
