% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.get_depthWeight.R
\name{get_depthWeight}
\alias{get_depthWeight}
\title{Calculate the depth weight for each sample or each experimental condition}
\usage{
get_depthWeight(sqlite_db, hugeData)
}
\arguments{
\item{sqlite_db}{A path to the SQLite database for InPAS, i.e. the output of
\code{\link[=setup_sqlitedb]{setup_sqlitedb()}}.}

\item{hugeData}{A logical(1), indicating whether it is huge data}
}
\value{
A named numeric vector containing depth weight for each sample for
non-hugeData, or depth weight for each condition if hugeData.
}
\description{
Calculate the depth weight for each sample of non-hugeData or each
experimental condition for hugeData: depth/mean(depth)
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
