% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.get_totalCov.R
\name{get_totalCov}
\alias{get_totalCov}
\title{Calculate the total coverage}
\usage{
get_totalCov(sqlite_db, seqname, hugeData)
}
\arguments{
\item{sqlite_db}{A path to the SQLite database for InPAS, i.e. the output of
\code{\link[=setup_sqlitedb]{setup_sqlitedb()}}.}

\item{seqname}{A character(1), the chromosome/scaffold name}

\item{hugeData}{A logical(1), indicating whether it is huge data}
}
\value{
A list containing coverage data. For hugeData, coverage of samples
in each condition is merged chromosome by chromosome. For non-hugeData,
per-chromosome coverage of all samples are returned.
\describe{
\item{seqname}{chromosome/scaffold name}
\describe{
\item{tag1}{name tag for sample1}
\item{tag2}{name tag for sample2}
\item{tagN}{name tag for sampleN}
}
}
}
\description{
For hugeData, coverage of samples in each condition is merged chromosome
by chromosome. For non-hugeData, per-chromosome coverage of all samples
are returned.
}
\author{
Haibo Liu, Jianhong Ou
}
\keyword{internal}
