% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18.get_UTR3region.R
\name{get_UTR3region}
\alias{get_UTR3region}
\title{extract long and short 3UTR region}
\usage{
get_UTR3region(.grs)
}
\arguments{
\item{.grs}{output of \code{\link[=search_CPs]{search_CPs()}}}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object with short form and long 3' UTR forms
}
\description{
extract long and short 3UTR region
}
\author{
Jianhong Ou
}
\keyword{internal}
