% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.get_UTR3TotalCov.R
\name{get_UTR3TotalCov}
\alias{get_UTR3TotalCov}
\title{extract coverage of 3' UTR for CP sites prediction}
\usage{
get_UTR3TotalCov(
  utr3,
  totalCov,
  BPPARAM = NULL,
  gcCompensation = NA,
  mappabilityCompensation = NA,
  FFT = FALSE,
  fft.sm.power = 20
)
}
\arguments{
\item{utr3}{An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}. It must be the
output of \code{\link[=extract_UTR3Anno]{extract_UTR3Anno()}}}

\item{totalCov}{total coverage for each sample. It must be a list of the
output of \code{\link[=get_totalCov]{get_totalCov()}}}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply. It can be set to NULL or bpparam()}

\item{gcCompensation}{GC compensation vector. Not support yet.}

\item{mappabilityCompensation}{mappability compensation vector. Not support
yet.}

\item{FFT}{Use FFT smooth or not.}

\item{fft.sm.power}{the cut-off frequency of FFT smooth.}
}
\value{
A list containing total coverage for each 3' UTR.
\describe{
\item{seqname}{chromosome/scaffold name}
\describe{
\item{transcript1}{data matrix containing summarized View
for transcript1}
\item{transcript2}{data matrix containing summarized View
for transcript2}
\item{transcriptN}{data matrix containing summarized View
for transcriptN}
}
}
}
\description{
extract 3' UTR coverage from totalCov according to the
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object utr3.
}
\author{
Jianhong Ou
}
\keyword{internal}
