% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09.get_zScoreCutoff.R
\name{get_zScoreCutoff}
\alias{get_zScoreCutoff}
\title{Calculate local background cutoff value}
\usage{
get_zScoreCutoff(background, introns, totalCov, utr3, z = 2)
}
\arguments{
\item{background}{A character(1) vector, indicating how background coverage
is defined.}

\item{introns}{An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} for introns}

\item{totalCov}{total coverage, a list of output from \code{\link[=get_totalCov]{get_totalCov()}} for
each chromosome}

\item{utr3}{An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}, output of
\code{\link[=extract_UTR3Anno]{extract_UTR3Anno()}}}

\item{z}{Z score cutoff value}
}
\value{
A named numeric vector containing local background Z-score cutoff
values. The names are GRanges's name for 3' UTRs.
}
\description{
calculate local background z-score cutoff
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
