% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{getChr2Exclude}
\alias{getChr2Exclude}
\title{Get a globally-applied requirement for filtering scaffolds.}
\usage{
getChr2Exclude()
}
\description{
This function will get the default requirement of filtering scaffolds.
}
