% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getCov}
\title{extract coverage from bedgraph file}
\usage{
getCov(bedgraph, genome, seqLen, removeScaffolds = FALSE)
}
\arguments{
\item{bedgraph}{bedGraph file names}

\item{genome}{an object \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{seqLen}{a named vector containing lengths of each chromosome. output
from \code{\link[=seqLen]{seqLen()}}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}
}
\value{
an object of \link[S4Vectors:Rle-class]{S4Vectors::Rle} for a sample coverage
}
\description{
extract coverage from bedgraph file
}
\seealso{
\code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}
}
