% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov4HugeData.R
\name{getCov4HugeData}
\alias{getCov4HugeData}
\title{Process coverage files for hugData}
\usage{
getCov4HugeData(coverageFiles, sqlite_db, outdir)
}
\arguments{
\item{coverageFiles}{A vector of character(n), paths to R objects storing
coverage filenames and depth for each sample, which are output of
bedgraph2RleCov() of Bam2RleCov() with hugeData set to "TRUE"}

\item{tags}{unique sample name tags for each corresponding bedGraph file}
}
\value{
A list of coverage with tags as names
}
\description{
Process per-sample coverage files of an experiment into a list
per chromosome coverage file and depth per sample
}
