% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov4SmallData.R
\name{getCov4SmallData}
\alias{getCov4SmallData}
\title{Get coverage for a small-scale experiment}
\usage{
getCov4SmallData(
  bedgraphs,
  tags,
  genome,
  removeScaffolds = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{bedgraphs}{a vector of character, containing paths to bedGraph files,
absolute paths or path relative to the current working directory}

\item{tags}{unique sample name tags matching each bedGraph file}

\item{genome}{an object \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply. It can be set to NULL or bpparam()}

\item{hugeData}{a vector of logical(1), indicating whether the data is huge.
For experiments with more than 3 samples, hugeData = TRUE should be
considered.}

\item{outdir}{a directory with write permission for storing intermediate
coverage data. This is required for hughe data, i.e., when hugeData is set
to TRUE}
}
\value{
a list of coverage files or objects of \link[S4Vectors:Rle-class]{S4Vectors::Rle}
}
\description{
Get coverage for a small-scale experiment
}
