% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverage.R
\name{getCovFileList}
\alias{getCovFileList}
\title{Combine coverage files}
\usage{
getCovFileList(coverageFiles, tags)
}
\arguments{
\item{coverageFiles}{paths to coverage files, absolute paths or path relative
to the current working directory}

\item{tags}{unique sample name tags for each corresponding bedGraph file}
}
\value{
A list of coverage with tags as names
}
\description{
Combine per-sample coverage files of an experiment into a list
}
