% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{getInPASEnsDb}
\alias{getInPASEnsDb}
\title{Get the globally defined EnsDb.}
\usage{
getInPASEnsDb()
}
\value{
An object of \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}
}
\description{
Get the globally defined EnsDb.
}
