% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{getInPASOutputDirectory}
\alias{getInPASOutputDirectory}
\title{Get the path to a output directory for InPAS analysis}
\usage{
getInPASOutputDirectory()
}
\value{
a normalized path to a output directory for InPAS analysis
}
\description{
Get the path to a output directory for InPAS analysis
}
