% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.setup_sqlitedb.R
\name{getInPASSQLiteDb}
\alias{getInPASSQLiteDb}
\title{Get the path to an SQLite database}
\usage{
getInPASSQLiteDb()
}
\value{
A path to an SQLite database
}
\description{
Get the path to an SQLite database
}
