% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{getInPASTxDb}
\alias{getInPASTxDb}
\title{Get the globally defined TxDb.}
\usage{
getInPASTxDb()
}
\value{
An object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}
}
\description{
Get the globally defined TxDb.
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
addInPASTxDb(TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene)
getInPASTxDb()
}
