% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.set_globals.R
\name{getLockName}
\alias{getLockName}
\title{Get the path to a file for locking the SQLite database}
\usage{
getLockName()
}
\value{
A path to a file for locking
}
\description{
Get the path to a file for locking the SQLite database
}
