% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionCoverage.R
\name{getRegionCoverage}
\alias{getRegionCoverage}
\title{calculate coverage for a given region}
\usage{
getRegionCoverage(chr, utr3.regions.chr, hugeData, coverage, phmm = FALSE)
}
\arguments{
\item{chr}{character(1), chromosome/scaffold name}

\item{utr3.regions.chr}{An object of \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList},
with a GRanges for a single seqname (chromosome/scaffold). It is an
output from the \code{\link[=get3UTRCDSRegionsPerSeqname]{get3UTRCDSRegionsPerSeqname()}}}

\item{hugeData}{is it a huge dataset?}

\item{coverage}{output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}, a list of coverage file names
or a list of coverage objects}

\item{phmm}{prepare data for singleSample analysis?}
}
\value{
GRanges with coverage data
}
\description{
calculate coverage for a given region
}
\keyword{internal}
