% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverage.R
\name{getSingleSampleCov}
\alias{getSingleSampleCov}
\title{extract coverage from a bedgraph file}
\usage{
getSingleSampleCov(
  bedgraph,
  genome,
  removeScaffolds = FALSE,
  hugeData = TRUE,
  BPPARAM = NULL,
  outdir = NULL
)
}
\arguments{
\item{bedgraph}{A bedGraph file name with its accessible path}

\item{genome}{an object \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{hugeData}{is this dataset consume too much memory? if it is TRUE, the
coverage will be saved into tempfiles.}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply. It can be set to NULL or bpparam()}

\item{outdir}{a directory with write permission for storing intermediate
coverage data. This is required for huge data, i.e., when hugeData is set
to TRUE}
}
\value{
an object of \link[S4Vectors:Rle-class]{S4Vectors::Rle} for a sample coverage
}
\description{
extract coverage from a bedgraph file
}
\examples{
  library(BSgenome.Mmusculus.UCSC.mm10)
  bedgraph <- system.file("extdata",
    "Baf3.extract.bedgraph",
    package = "InPAS"
  )
  genome <- BSgenome.Mmusculus.UCSC.mm10
  coverage <- getSingleSampleCov(bedgraph,
    genome,
    removeScaffolds = TRUE,
    hugeData = FALSE,
    outdir = tempdir()
  )
  coverage <- getSingleSampleCov(bedgraph,
    genome,
    removeScaffolds = TRUE,
    hugeData = TRUE,
    BPPARAM = BiocParallel::bpparam(),
    outdir = tempdir()
  )
}
