% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUTR3region.R
\name{getUTR3region}
\alias{getUTR3region}
\title{extract long and short 3UTR region}
\usage{
getUTR3region(.grs)
}
\arguments{
\item{.grs}{output of \code{\link[=CPsites]{CPsites()}}}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object with short form and long 3' UTR forms
}
\description{
extract long and short 3UTR region
}
\keyword{internal}
