% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate3UTRUsage.R
\name{integrate3UTRUsage}
\alias{integrate3UTRUsage}
\title{Assemble chromosome-wise 3'UTR coverage data into a GRanges object}
\usage{
integrate3UTRUsage(UTR3CoverageFiles)
}
\arguments{
\item{UTR3CoverageFiles}{paths to files storing chromosome-wise 3'UTR
coverage data for each sample}
}
\value{
An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
}
\description{
Assemble chromosome-wise 3'UTR coverage data for each sample into a GRanges
object
}
