% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateCPsites.R
\name{integrateCPsites}
\alias{integrateCPsites}
\title{Combine and format CPsites}
\usage{
integrateCPsites(shorten_UTR_estimation_files, utr3)
}
\arguments{
\item{shorten_UTR_estimation_files}{A vector of character, containing paths to all
the outputs of \code{\link[=estimateCPsites]{estimateCPsites()}} of an experiment. absolute paths or
paths relative to current work directory}

\item{utr3}{An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}, output from the
\code{\link[=utr3Annotation]{utr3Annotation()}}}
}
\value{
An object of \code{\link[GenomicRanges]{GRanges}}
}
\description{
Combine and format CPsites
}
