% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covThreshold.R
\name{intronRegion}
\alias{intronRegion}
\title{Helper function to get intronic regions}
\usage{
intronRegion(TxDb)
}
\arguments{
\item{TxDb}{An object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}
}
\value{
GRanges of intronic regions not overlapping exons
}
\description{
Helper function to get intronic regions
}
\keyword{internal}
