% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaAnalysis}
\alias{limmaAnalysis}
\title{use limma to analyze the PDUI}
\usage{
limmaAnalysis(UTR3eset, design, contrast.matrix, coef = 1, robust = FALSE, ...)
}
\arguments{
\item{UTR3eset}{an \linkS4class{UTR3eSet} object}

\item{design}{the design matrix of the experiment, with rows corresponding to
arrays and columns to coefficients to be estimated. Defaults to the unit
vector meaning that the arrays are treated as replicates. see
\code{\link[stats:model.matrix]{stats::model.matrix()}}}

\item{contrast.matrix}{numeric matrix with rows corresponding to coefficients
in fit and columns containing contrasts. May be a vector if there is only
one contrast. see \code{\link[limma:makeContrasts]{limma::makeContrasts()}}}

\item{coef}{column number or column name specifying which coefficient or
contrast of the linear model is of interest. see more \code{\link[limma:toptable]{limma::topTable()}}.
default value: 1}

\item{robust}{logical, should the estimation of the empirical Bayes prior
parameters be robustified against outlier sample variances?}

\item{...}{other arguments are passed to lmFit}
}
\value{
fit results of eBayes by limma. It is an object of class
\link[limma:marraylm]{limma::MArrayLM} containing everything found in fit. see
\code{\link[limma:ebayes]{limma::eBayes()}}
}
\description{
use limma to analyze the PDUI
}
\examples{
library(limma)
path <- system.file("extdata", package = "InPAS")
load(file.path(path, "eset.MAQC.rda"))
tags <- colnames(eset$PDUI.log2)
g <- factor(gsub("\\\\..*$", "", tags))
design <- model.matrix(~ -1 + g)
colnames(design) <- c("Brain", "UHR")
contrast.matrix <- makeContrasts(contrasts = "Brain-UHR", levels = design)
res <- limmaAnalysis(eset, design, contrast.matrix)
head(res)
}
\seealso{
\code{\link[=singleSampleAnalysis]{singleSampleAnalysis()}}, \code{\link[=singleGroupAnalysis]{singleGroupAnalysis()}}, \code{\link[=fisherExactTest]{fisherExactTest()}}
}
