% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{mapComp}
\alias{mapComp}
\title{Calculate weights for mappability-base coverage correction}
\usage{
mapComp(mi)
}
\arguments{
\item{mi}{A numeric vector of mappability along per
chromosome/scaffold}
}
\value{
A numeric vector of weights for mappability-based
coverage correction
}
\description{
Calculate weights for mappability-base coverage correction
}
\keyword{internal}
