% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCoverage.R
\name{mergeCoverage}
\alias{mergeCoverage}
\title{merge read coverage from different samples}
\usage{
mergeCoverage(
  coverage,
  groupList,
  genome,
  removeScaffolds = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{coverage}{coverage for each sample, output of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply}
}
\value{
A list, coverage for a group of samples
}
\description{
merge read coverage from different samples
}
\keyword{internal}
