% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalSegmentation.R
\name{optimalSegmentation}
\alias{optimalSegmentation}
\title{calculate MSE}
\usage{
optimalSegmentation(
  .ele,
  search_point_START,
  search_point_END,
  n = 1,
  savedID = NA
)
}
\arguments{
\item{.ele}{3' UTR coverage}

\item{search_point_START}{start position to calculate}

\item{search_point_END}{end position to calculate}

\item{n}{the length of output. If set to -1, return all candidate CP sites}

\item{savedID}{the proximal CPsites}
}
\value{
a list of MSE and candidate APA sites
}
\description{
calculate MSE values for each assumed CP site
}
\keyword{internal}
