% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTxDb.R
\name{parseTxDb}
\alias{parseTxDb}
\title{Extract gene models from a TxDb object}
\usage{
parseTxDb(TxDb = NULL, edb = NULL, removeScaffolds = FALSE)
}
\arguments{
\item{TxDb}{An object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}

\item{edb}{An object of \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be
removed from the genome. If you use a TxDb containing alternative
scaffolds, you'd better to remove the scaffolds.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object for gene models
}
\description{
Extract gene models from a TxDb object.
}
\details{
A good practice is to perform read alignment using a
reference genome from Ensembl/GenCode including only the primary
assembly and build a TxDb using the GTF/GFF files downloaded
from the same source as the reference genome, such as
BioMart/Ensembl/GenCode. For instruction, see Vignette of the
GenomicFeatures. The UCSC reference genomes and their
annotation can be very cubersome.
}
\examples{
library("EnsDb.Hsapiens.v86")
library("GenomicFeatures")
samplefile <- system.file("extdata",
  "hg19_knownGene_sample.sqlite",
  package = "GenomicFeatures"
)
TxDb <- loadDb(samplefile)
edb <- EnsDb.Hsapiens.v86

parsed_Txdb <- parseTxDb(TxDb, edb, removeScaffolds = TRUE)
}
\author{
Haibo Liu
}
