% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAscore.R
\name{PAscore}
\alias{PAscore}
\title{calculate the CP score}
\usage{
PAscore(seqname, pos, str, idx, PWM, genome, ups = 50, dws = 50)
}
\arguments{
\item{seqname}{sequence names}

\item{pos}{genomic positions}

\item{str}{strands}

\item{idx}{offset position}

\item{PWM}{polyA position weight matrix}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{ups}{upstream base}

\item{dws}{downstream base}
}
\value{
A list containing offset positions after filtering
}
\description{
calculate the CP score by using PWM of polyadenylation signal
}
\seealso{
\code{\link[=PAscore2]{PAscore2()}}
}
\keyword{internal}
