% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAscore2.R
\name{PAscore2}
\alias{PAscore2}
\title{calculate the CP score}
\usage{
PAscore2(seqname, pos, str, idx, idx.gp, genome, classifier, classifier_cutoff)
}
\arguments{
\item{seqname}{sequence names}

\item{pos}{genomic positions}

\item{str}{strands}

\item{idx}{offset position}

\item{idx.gp}{group number of the offset position}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{classifier}{a cleanUpdTSeq classifier}

\item{classifier_cutoff}{classifier cutoff value}
}
\value{
a data frame
}
\description{
calculate CP score by cleanUpdTSeq
}
\seealso{
\code{\link[=PAscore]{PAscore()}}
}
\keyword{internal}
