% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27.plot_utr3Usage.R
\name{plot_utr3Usage}
\alias{plot_utr3Usage}
\title{Visualize the dPDUI events using ggplot2}
\usage{
plot_utr3Usage(usage_data, vline_color = "purple", vline_type = "dashed")
}
\arguments{
\item{usage_data}{An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}, an output from
\code{\link[=get_usage4plot]{get_usage4plot()}}.}

\item{vline_color}{color for vertical line showing position of predicated
proximal CP site. Default, purple.}

\item{vline_type}{line type for vertical line showing position of predicated
proximal CP site. Default, dashed. See \href{https://cran.r-project.org/web/packages/ggplot2/vignettes/ggplot2-specs.html}{ggplot2 linetype}.}
}
\value{
A ggplot object for refined plotting
}
\description{
Visualize the dPDUI events by plotting the MSE, and total coverage per group
along 3' UTR regions with dPDUI using \code{\link[ggplot2:geom_path]{ggplot2::geom_line ()}}.
}
\seealso{
For example, see \code{\link[=get_usage4plot]{get_usage4plot()}}.
}
\author{
Haibo Liu
}
