% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4CPsiteSearch.R
\name{prepare4CPsiteSearch}
\alias{prepare4CPsiteSearch}
\title{prepare data for predicting cleavage and polyadenylation (CP) sites}
\usage{
prepare4CPsiteSearch(
  coverage,
  groupList = NULL,
  genome,
  utr3,
  background = c("same_as_long_coverage_threshold", "1K", "5K", "10K", "50K"),
  TxDb = NA,
  removeScaffolds = FALSE,
  BPPARAM = NULL,
  tmpfolder = NULL,
  silence = TRUE
)
}
\arguments{
\item{coverage}{coverage for each sample, output from
\code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}

\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{utr3}{output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{background}{the range for calculating cutoff threshold of local
background}

\item{TxDb}{an object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be
removed from the genome. If you use a TxDb containing alternative
scaffolds, you'd better to remove the scaffolds.}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply.}

\item{tmpfolder}{temp folder which could be used to save and reload the
analysis data for resuming analysis}

\item{silence}{report progress or not. By default it doesn't report progress.}
}
\value{
An object of \code{\link[GenomicRanges]{GRanges}}
}
\description{
prepare data for predicting cleavage and polyadenylation (CP) sites
}
\examples{
if (interactive()) {
  library(BSgenome.Mmusculus.UCSC.mm10)
  bedgraphs <- system.file("extdata",
    "Baf3.extract.bedgraph",
    package = "InPAS"
  )
  data(utr3.mm10)
  tags <- "Baf3"
  genome <- BSgenome.Mmusculus.UCSC.mm10
  coverage <- coverageFromBedGraph(bedgraphs,
    tags, genome,
    removeScaffolds = FALSE,
    hugeData = FALSE
  )
  CP <- CPsites(
    coverage = coverage,
    groupList = tags,
    genome = genome,
    utr3 = utr3.mm10,
    coverage_threshold = 5,
    long_coverage_threshold = 5
  )
}
}
\references{
Cheung MS, Down TA, Latorre I, Ahringer J. Systematic bias in
high-throughput sequencing data and its correction by BEADS. Nucleic Acids
Res. 2011 Aug;39(15):e103. doi: 10.1093/nar/gkr425. Mappability could be
calculated by \href{http://algorithms.cnag.cat/wiki/Man:gem-mappability}{GEM}
Derrien T, Estelle J, Marco Sola S, Knowles DG, Raineri E, Guigo R, Ribeca
P. Fast computation and applications of genome mappability. PLoS One.
2012;7(1):e30377. doi: 10.1371/journal.pone.0030377.
}
