% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_depthNcov.R
\name{process_depthNcov}
\alias{process_depthNcov}
\title{Format depth and coverage data}
\usage{
process_depthNcov(coverage)
}
\arguments{
\item{coverage}{A vector of character(n) containing paths to RDS files
storing Rle instances and sample total read depth for hugeData, or
a list containing Rle instances and sample total read depth}

\item{outdir}{A path with write permission. If it is NULL, the output will
be written to the same directory as the input coverage files for hugeData}
}
\value{

}
\description{
Glean sample depth data together and convert sample-wise coverage data to
chromosome-wise coverage data
}
