% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximalAdj.R
\name{proximalAdj}
\alias{proximalAdj}
\title{adjust the proximal CP sites}
\usage{
proximalAdj(
  CPs,
  MINSIZE,
  PolyA_PWM,
  genome,
  classifier,
  classifier_cutoff,
  shift_range,
  search_point_START,
  step = 1
)
}
\arguments{
\item{CPs}{the outputs of \code{\link[=searchProximalCPs]{searchProximalCPs()}}}

\item{MINSIZE}{min size for short from}

\item{PolyA_PWM}{PolyA position weight matrix}

\item{genome}{a \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object}

\item{classifier}{cleanUpdTSeq classifier}

\item{classifier_cutoff}{cutoff value of the classifier}

\item{shift_range}{the searching range for the better CP sites}

\item{search_point_START}{just in case there is no better CP sites}

\item{step}{adjust step, default 1, means adjust by each base by
cleanUpdTSeq.}
}
\value{
keep same as \code{\link[=searchProximalCPs]{searchProximalCPs()}}, which can be handled by
\code{\link[=polishCPs]{polishCPs()}}.
}
\description{
adjust the proximal CP sites by PolyA PWM and cleanUpdTSeq
}
\seealso{
\code{\link[=searchProximalCPs]{searchProximalCPs()}}, \code{\link[=polishCPs]{polishCPs()}}, \code{\link[=proximalAdjByPWM]{proximalAdjByPWM()}},
\code{\link[=proximalAdjByCleanUpdTSeq]{proximalAdjByCleanUpdTSeq()}}, \code{\link[=PAscore]{PAscore()}},\code{\link[=PAscore2]{PAscore2()}}
}
\keyword{internal}
