% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximalAdjByPWM.R
\name{proximalAdjByPWM}
\alias{proximalAdjByPWM}
\title{adjust the proximal CP sites by matching PWM}
\usage{
proximalAdjByPWM(
  idx,
  PolyA_PWM,
  seqnames,
  starts,
  strands,
  genome,
  shift_range,
  search_point_START
)
}
\arguments{
\item{idx}{the offset of positions of CP sites}

\item{PolyA_PWM}{polyA PWM}

\item{seqnames}{sequence names}

\item{starts}{start position in the genome}

\item{strands}{strands}

\item{genome}{an \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object}

\item{shift_range}{the shift range of PWM hits}

\item{search_point_START}{Not use}
}
\value{
the offset of positions of CP sites after filter
}
\description{
adjust the proximal CP sites by polyA Position Weight Matrix. It only need
the PWM to get match in upstream or downstream shift_range nr.
}
\details{
the hits is searched by \code{\link[Biostrings:matchPWM]{Biostrings::matchPWM()}} and the cutoff is
70\\%
}
\seealso{
\code{\link[=proximalAdjByCleanUpdTSeq]{proximalAdjByCleanUpdTSeq()}}, \code{\link[=PAscore]{PAscore()}}
}
\keyword{internal}
