% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm}
\alias{pwm}
\title{polyA motif}
\format{
A matrix with 6 columns, 4 rows. And rownames are "A", "C", "G",
and "T".
}
\usage{
pwm
}
\description{
A position weight matrix for polyA
}
\keyword{datasets}
