% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUTR3__UTR3.R
\name{removeUTR3__UTR3}
\alias{removeUTR3__UTR3}
\title{remove the converging candidates 3' UTRs LIKE UTR3___UTR3}
\usage{
removeUTR3__UTR3(x)
}
\arguments{
\item{x}{the distal 3UTR coverage}
}
\value{
the 3UTR coverage after removing the next 3UTR
}
\description{
some of the results is from connected two 3' UTRs. We want to remove them.
}
\details{
The algorithm need to be improved.
}
\keyword{internal}
