% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23.run_fisherExactTest.R
\name{run_fisherExactTest}
\alias{run_fisherExactTest}
\title{Run Fisher Exact Test for differential usage of 3' UTRs for a two-group experimental design}
\usage{
run_fisherExactTest(UTR3eset, gp1, gp2)
}
\arguments{
\item{UTR3eset}{An object of \linkS4class{UTR3eSet}, output of \code{\link[=get_UTR3eSet]{get_UTR3eSet()}}}

\item{gp1}{tag names of group 1}

\item{gp2}{tag names of group 2}
}
\value{
a matrix of test results
}
\description{
Run Fisher Exact Test for differential usage of 3' UTRs for a two-group experimental design
}
\seealso{
\code{\link[=run_singleSampleAnalysis]{run_singleSampleAnalysis()}} for a single-sample APA
analysis,\code{\link[=run_singleGroupAnalysis]{run_singleGroupAnalysis()}} for a single-group sample APA analysis,
\code{\link[=run_limmaAnalysis]{run_limmaAnalysis()}} for limma-based APA analysis of complex experimental
design
}
\author{
Jianhong Ou
}
\keyword{internal}
