% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23.run_singleGroupAnalysis.R
\name{run_singleGroupAnalysis}
\alias{run_singleGroupAnalysis}
\title{do analysis for single group samples}
\usage{
run_singleGroupAnalysis(UTR3eset)
}
\arguments{
\item{UTR3eset}{An object of \linkS4class{UTR3eSet}, output of \code{\link[=get_UTR3eSet]{get_UTR3eSet()}}}
}
\value{
a matrix of test results
}
\description{
do analysis for single group samples by ANOVA test
}
\examples{
path <- system.file("extdata", package = "InPAS")
load(file.path(path, "eset.MAQC.rda"))
res <- InPAS:::run_singleGroupAnalysis(eset)
}
\author{
Jianhong Ou
}
\keyword{internal}
