% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23.run_singleSampleAnalysis.R
\name{run_singleSampleAnalysis}
\alias{run_singleSampleAnalysis}
\title{do APA analysis for a single sample}
\usage{
run_singleSampleAnalysis(UTR3eset)
}
\arguments{
\item{UTR3eset}{the output of \code{\link[=get_UTR3eSet]{get_UTR3eSet()}}}
}
\value{
a matrix containing test results
}
\description{
do APA event analysis for a single sample Using Poisson Hidden
Markov models
}
\details{
the test will be performed by comparing a two-state versus an
one-state Poisson Hidden Markov models.
}
\examples{
path <- system.file("extdata", package = "InPAS")
load(file.path(path, "eset.MAQC.rda"))
res <- InPAS:::run_singleSampleAnalysis(eset)
}
\seealso{
\linkS4class{UTR3eSet}, \code{\link[=get_UTR3eSet]{get_UTR3eSet()}}, \code{\link[depmixS4:depmix]{depmixS4::depmix()}}
}
\author{
Jianhong Ou
}
\keyword{internal}
