% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.search_proximalCPs.R
\name{search_proximalCPs}
\alias{search_proximalCPs}
\title{search proximal CPsites}
\usage{
search_proximalCPs(
  CPs,
  curr_UTR,
  window_size,
  MINSIZE,
  cutEnd,
  search_point_START,
  search_point_END,
  two_way = FALSE
)
}
\arguments{
\item{CPs}{output from \code{\link[=search_distalCPs]{search_distalCPs()}} or \code{\link[=adjust_distalCPs]{adjust_distalCPs()}}}

\item{curr_UTR}{GRanges for current 3' UTR}

\item{window_size}{window size}

\item{MINSIZE}{MINSIZE for short form}

\item{cutEnd}{a numeric(1) between 0 and 1. The percentage of nucleotides
should be removed from the end before search, 0.1 means 10 percent.}

\item{search_point_START}{start point for searching}

\item{search_point_END}{end point for searching}

\item{two_way}{Search the proximal site from both direction or not.}
}
\value{
a list
}
\description{
search proximal CPsites
}
\seealso{
\code{\link[=adjust_proximalCPs]{adjust_proximalCPs()}}, \code{\link[=polish_CPs]{polish_CPs()}},\code{\link[=adjust_proximalCPsByPWM]{adjust_proximalCPsByPWM()}},
\code{\link[=adjust_proximalCPsByNBC]{adjust_proximalCPsByNBC()}}, \code{\link[=get_PAscore]{get_PAscore()}}, \code{\link[=get_PAscore2]{get_PAscore2()}}
}
\author{
Jianhong Ou
}
\keyword{internal}
