% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLen.R
\name{seqLen}
\alias{seqLen}
\title{Get sequence lengths}
\usage{
seqLen(genome, removeScaffolds = FALSE)
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}
}
\value{
A numeric vector containing lengths per seqnames
}
\description{
Get sequence lengths from a \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object
}
\seealso{
\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo}
}
\keyword{internal}
