% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sqlitedb.R
\name{setup_sqlitedb}
\alias{setup_sqlitedb}
\title{Create an SQLite database for storing metadata and paths to coverage files}
\usage{
setup_sqlitedb(metadata, outdir)
}
\arguments{
\item{metadata}{A path to a tab-delimited file, with columns "tag", "group",
and "bedgraph_file", storing a unique name tag for each sample, a condition
name for each sample, such as "treatment" and "control", and a path to the
bedgraph file for each sample}

\item{outdir}{A character(1), a path with write permission for storing the
SQLite database}
}
\value{
A character(1), the path to the SQLite database
}
\description{
Create an SQLite database with two tables, metadata and chromosome_coverage,
for storing metadata (sample tag, condition, paths to begraph files, and sample
total read coverage), and paths to chromosome-oriented coverage files,
respectively
}
