% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGroupAnalysis.R
\name{singleGroupAnalysis}
\alias{singleGroupAnalysis}
\title{do analysis for single group samples}
\usage{
singleGroupAnalysis(UTR3eset)
}
\arguments{
\item{UTR3eset}{output from \code{\link[=getUTR3eSet]{getUTR3eSet()}}}
}
\value{
a matrix of test results
}
\description{
do analysis for single group samples by ANOVA test
}
\examples{
path <- system.file("extdata", package = "InPAS")
load(file.path(path, "eset.MAQC.rda"))
res <- singleGroupAnalysis(eset)
}
