% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGroupAnalyze.R
\name{singleGroupAnalyze}
\alias{singleGroupAnalyze}
\title{do analysis for single group samples}
\usage{
singleGroupAnalyze(UTR3eset)
}
\arguments{
\item{UTR3eset}{output from [getUTR3eSet()]}
}
\value{
a matrix of test results
}
\description{
do analysis for single group samples by anova test
}
\examples{
path <- file.path(find.package("InPAS"), "extdata")
load(file.path(path, "eset.MAQC.rda"))
res <- singleGroupAnalyze(eset)
}
