% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleAnalysis.R
\name{singleSampleAnalysis}
\alias{singleSampleAnalysis}
\title{do APA analysis for a single sample}
\usage{
singleSampleAnalysis(UTR3eset)
}
\arguments{
\item{UTR3eset}{the output of \code{\link[=getUTR3eSet]{getUTR3eSet()}}}
}
\value{
a matrix containing test results
}
\description{
do APA event analysis for a single sample Using Poisson Hidden
Markov models
}
\details{
the test will be performed by comparing a two-state versus an
one-state Poisson Hidden Markov models.
}
\examples{
path <- system.file("extdata", package = "InPAS")
load(file.path(path, "eset.MAQC.rda"))
res <- singleSampleAnalysis(eset)
}
\seealso{
\linkS4class{UTR3eSet}, \code{\link[=getUTR3eSet]{getUTR3eSet()}}, \code{\link[depmixS4:depmix]{depmixS4::depmix()}}
}
