% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleAnalyze.R
\name{singleSampleAnalyze}
\alias{singleSampleAnalyze}
\title{do APA analysis for a single sample}
\usage{
singleSampleAnalyze(UTR3eset)
}
\arguments{
\item{UTR3eset}{the output of [getUTR3eSet()]}
}
\value{
a matrix containing test results
}
\description{
do APA event analysis for a single sample Using Poisson Hidden Markov models
}
\details{
the test will be performed by comparing a two-state versus
an one-state Poisson Hidden Markov models.
}
\examples{
1
path <- file.path(find.package("InPAS"), "extdata")
load(file.path(path, "eset.MAQC.rda"))
res <- singleSampleAnalyze(eset)
}
\seealso{
[UTR3eSet-class], [getUTR3eSet()], [depmixS4::depmix()]
}
