% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortGR.R
\name{sortGR}
\alias{sortGR}
\title{sort GRanges}
\usage{
sortGR(.ele)
}
\arguments{
\item{.ele}{an object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}}
}
\value{
a sorted \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object
}
\description{
sort GRanges on the positive/negative strand by start/end position in
ascending/descending order, respectively
}
\keyword{internal}
