% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalCoverage.R
\name{totalCoverage}
\alias{totalCoverage}
\title{calculate the total coverage}
\usage{
totalCoverage(
  coverage,
  genome,
  hugeData = FALSE,
  removeScaffolds = FALSE,
  groupList = NULL
)
}
\arguments{
\item{coverage}{coverage for each sample, outputs of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{hugeData}{hugeData or not}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}
}
\value{
a coverage list
}
\description{
For huge dataset, it will read in the coverage from temp files and merge them
by groups
}
\keyword{internal}
