% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSeqnames.R
\name{trimSeqnames}
\alias{trimSeqnames}
\title{filter the sequence names from a BSgenome object}
\usage{
trimSeqnames(genome, removeScaffolds = FALSE)
}
\arguments{
\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}
}
\value{
an character vector of filtered seqnames
}
\description{
only chromosome-level seqnames are kept
}
\keyword{internal}
