% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage4plot.R
\name{usage4plot}
\alias{usage4plot}
\title{prepare coverage data and fitting data for plot}
\usage{
usage4plot(gr, coverage, proximalSites, genome, groupList = NULL)
}
\arguments{
\item{gr}{an object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}}

\item{coverage}{an object of \link[S4Vectors:Rle-class]{S4Vectors::Rle}}

\item{proximalSites}{proximal CP sites}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{groupList}{a list of grouped sample tag names, with the group names as
the list's name, such as list(groupA = c("sample_1", "sample_2",
"sample_3"), groupB = c("sample_4", "sample_5", "sample_6"))}
}
\value{
an object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} with metadata:
\item{dat}{matrix, first column is the fit data, the other columns are
coverage data for each sample}
\item{offset}{offset from the start of 3UTR}
}
\description{
prepare coverage data and fitting data for plot
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
path <- system.file("extdata", package = "InPAS")
bedgraphs <- c(
  file.path(path, "Baf3.extract.bedgraph"),
  file.path(path, "UM15.extract.bedgraph")
)
coverage <- coverageFromBedGraph(
  bedgraphs,
  tags = c("Baf3", "UM15"),
  genome = Mmusculus,
  removeScaffolds = TRUE,
  hugeData = FALSE
)
gr <- GRanges("chr6", IRanges(128846245, 128850081),
  strand = "-"
)
dat <- usage4plot(gr, coverage,
  proximalSites = 128849148, Mmusculus
)
data <- dat$dat[[1]]
op <- par(mfrow = c(3, 1))
plot(data[, 1],
  type = "l",
  xlab = "",
  ylab = "The fitted value"
)
abline(v = dat$offset)
plot(data[, 2],
  type = "l",
  xlab = "",
  ylab = "Baf3"
)
plot(data[, 3],
  type = "l",
  xlab = "",
  ylab = "UM15"
)
par(op)
}
