% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr3Annotation.R
\name{utr3Annotation}
\alias{utr3Annotation}
\title{extract 3' UTR information from a \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object}
\usage{
utr3Annotation(
  TxDb = NULL,
  edb = NULL,
  removeScaffolds = FALSE,
  MAX_EXONS_GAP = 10000
)
}
\arguments{
\item{TxDb}{an object of \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb}}

\item{edb}{An object of \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}}

\item{removeScaffolds}{logical(1), whether the scaffolds should be removed
from the genome}

\item{MAX_EXONS_GAP}{maximal gap sizes for distal CP site to downstream exons}
}
\value{
an object of GRanges with 7 metadata columns:
feature (utr3, next.exon.gap, CDS), annotatedProximalCP (unknown,
proximalCP_coordinate), exon (transcript id_index), transcript, gene
(entrez_id), symbol, truncated (logical).
}
\description{
extract 3' UTR information from a \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object. The
3'UTR is defined as the last 3'UTR fragment for each transcript and it will
be cut if there is any overlaps with other exons.
}
\details{
A good practice is to perform read alignment using a reference
genome from Ensembl/GenCode including only the primary assembly and build a
TxDb using the GTF/GFF files downloaded from the same source as the
reference genome, such as BioMart/Ensembl/GenCode. For instruction, see
Vignette of the GenomicFeatures. The UCSC reference genomes and their
annotation can be very cubersome.
}
\examples{
library("EnsDb.Hsapiens.v86")
library("GenomicFeatures")
samplefile <- system.file("extdata",
  "hg19_knownGene_sample.sqlite",
  package = "GenomicFeatures"
)
TxDb <- loadDb(samplefile)
edb <- EnsDb.Hsapiens.v86
utr3Annotation(TxDb, edb,
  removeScaffolds = TRUE,
  MAX_EXONS_GAP = 10000
)
}
\author{
Jianhong Ou, Haibo Liu
}
