% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20.UTR3eSet-class.R
\docType{class}
\name{UTR3eSet-class}
\alias{UTR3eSet-class}
\alias{$,UTR3eSet-method}
\alias{$<-,UTR3eSet-method}
\alias{coerce,UTR3eSet,ExpressionSet-method}
\alias{coerce,UTR3eSet,GRanges-method}
\alias{show,UTR3eSet-method}
\title{UTR3eSet-class and its methods}
\description{
An object of class \linkS4class{UTR3eSet} representing the results of 3' UTR usage;
methods for constructing, showing, getting and setting attributes of objects;
methods for coercing object of other class to \linkS4class{UTR3eSet} objects.
}
\section{Objects from the Class}{

Objects can be created by calls of the form \code{new("UTR3eSet", ...)}

 Objects can be created by calls of the form
\code{new("UTR3eSet", ...)}.
}

\section{Slots}{

\describe{
\item{\code{usage}}{Object of class \code{"GRanges"}}
\item{\code{PDUI}}{Object of class \code{"matrix"}}
\item{\code{PDUI.log2}}{Object of class \code{"matrix"}}
\item{\code{short}}{Object of class \code{"matrix"}}
\item{\code{long}}{Object of class \code{"matrix"}}
\item{\code{signals}}{Object of class \code{"list"}}
\item{\code{testRes}}{Object of class \code{"matrix"}}
}
}

\section{UTR3eSet-class methods}{

\describe{
\item{$}{\code{signature(x = "UTR3eSet")}: ... }
\item{$<-}{\code{signature(x = "UTR3eSet")}: ... }
\item{coerce}{\code{signature(from = "UTR3eSet", to = "ExpressionSet")}: ... }
\item{coerce}{\code{signature(from = "UTR3eSet", to = "GRanges")}: ... }
\item{show}{\code{signature(object = "UTR3eSet")}: ... }
}
}

\seealso{
\code{GRanges}
}
\author{
Jianhong Ou
}
