% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3usage.R
\name{UTR3usage}
\alias{UTR3usage}
\title{calculate the usage of long and short forms of UTR3}
\usage{
UTR3usage(CPsites, coverage, hugeData, BPPARAM = NULL, phmm = FALSE)
}
\arguments{
\item{CPsites}{outputs of \code{\link[=CPsites]{CPsites()}}}

\item{coverage}{coverage for each sample, outputs of \code{\link[=coverageFromBedGraph]{coverageFromBedGraph()}}}

\item{hugeData}{is this dataset consume too much memory? if it is TRUE, the
coverage will be saved into tempfiles.}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply.}

\item{phmm}{logical(1), should data be prepared for singleSample analysis? By
default, FALSE}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object
}
\description{
calculate the usage of long and short forms of UTR3 for the results of
\code{\link[=CPsites]{CPsites()}}
}
\seealso{
\code{\link[=CPsites]{CPsites()}}
}
\keyword{internal}
