% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{valley}
\alias{valley}
\title{find the local minimal mean squared error (MSE)}
\usage{
valley(x, ss, se, n = 1, savedID = NA, filterByPval = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{ss}{start searching position}

\item{se}{end searching position}

\item{n}{the length of output. If n=-1, output all the local minimal MSE
positions.}

\item{savedID}{saved positions of local minimal MSE}

\item{filterByPval}{logical. Filter the positions by p value or not.}
}
\value{
a numeric vector, position list.
}
\description{
for a giving numeric vectors, calculate the top N local minimal mean squared
error. The sites of local minimal MSE are saved if they are in the range of
searching start site to searching end site
}
\keyword{internal}
