% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScoreThreshold.R
\name{zScoreThreshold}
\alias{zScoreThreshold}
\title{Calculate local background cutoff value}
\usage{
zScoreThreshold(background, introns, totalCov, utr3, z = 2)
}
\arguments{
\item{background}{character(1), indicating how background coverage is defined}

\item{introns}{A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object for introns}

\item{totalCov}{total coverage, output from \code{\link[=totalCoverage]{totalCoverage()}}}

\item{utr3}{output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{z}{z score cutoff value}
}
\value{
a numeric vector
}
\description{
calculate local background cutoff value based on z-score
}
\keyword{internal}
