% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.parse_TxDb.R
\name{assign_feature}
\alias{assign_feature}
\title{Helper function to label the last component of a genomic feature for each
transcript}
\usage{
assign_feature(gr, feature_alt = "utr3")
}
\arguments{
\item{gr}{A tibble converted from an object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}}

\item{feature_alt}{A character(1) vector, specifying the type of genomic
features, such as "CDS", "exon", "utr3", "utr5".}
}
\value{
An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
}
\description{
Helper function to label the last component of a genomic feature for each
transcript
}
\author{
Haibo Liu
}
\keyword{internal}
