% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.get_UTR3TotalCov.R
\name{compensation}
\alias{compensation}
\title{Compensate the coverage with GC-content or mappability}
\usage{
compensation(view, comp, start, end)
}
\arguments{
\item{view}{A list of view object}

\item{comp}{A numeric vector of weight for GC composition or mappability}

\item{start}{An integer vector, starting coordinates}

\item{end}{An integer vector, end coordinates}
}
\value{
a list of GC composition or mappability corrected coverage
}
\description{
Compensate the coverage with GC-content or mappability
}
\author{
Jianhong Ou
}
\keyword{internal}
