% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covThreshold.R
\name{covThreshold}
\alias{covThreshold}
\title{calculate the cutoff threshold of coverage}
\usage{
covThreshold(coverage, genome, txdb, utr3, chr = "chr1", hugeData, groupList)
}
\arguments{
\item{coverage}{Coverage for each sample, output of [coverageFromBedGraph()]}

\item{genome}{An object of [BSgenome::BSgenome-class]}

\item{txdb}{An object of [GenomicFeatures::TxDb-class]}

\item{utr3}{Output of [utr3Annotation()]}

\item{chr}{Chromosome to be used for calculation, default is "chr1"}

\item{hugeData}{Is this dataset consume too much memory? if it is 
TRUE, the coverage will be saved into tempfiles.}

\item{groupList}{Group list of tag names}
}
\value{
A numeric vector
}
\description{
calculate the cutoff threshold of coverage for long and short isoforms
}
\keyword{internal}
