% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthWeight.R
\name{depthWeight}
\alias{depthWeight}
\title{calculate the depth weight for each example}
\usage{
depthWeight(coverage, hugeData, groupList = NULL)
}
\arguments{
\item{coverage}{a list. output of [coverageFromBedGraph()]}

\item{hugeData}{is it a huge dataset?}

\item{groupList}{group list for huge dataset}
}
\value{
a numeric vector with depth weight
}
\description{
calculate the depth weight for each example: depth/mean(depth)
}
\keyword{internal}
