% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCPsites.R
\name{estimateCPsites}
\alias{estimateCPsites}
\title{Estimate the CP sites}
\usage{
estimateCPsites(
  chr.cov,
  utr3,
  MINSIZE,
  window_size,
  search_point_START,
  search_point_END,
  cutStart,
  cutEnd,
  adjust_distal_polyA_end,
  background,
  z2s,
  coverage_threshold,
  long_coverage_threshold,
  PolyA_PWM,
  classifier,
  classifier_cutoff,
  shift_range,
  depth.weight,
  genome,
  step = 1,
  two_way = FALSE,
  tmpfolder = NULL,
  silence = TRUE
)
}
\arguments{
\item{chr.cov}{Coverage list for one chromosome}

\item{utr3}{output of \code{\link[=utr3Annotation]{utr3Annotation()}}}

\item{MINSIZE}{min size of short form}

\item{window_size}{window size}

\item{search_point_START}{search start point}

\item{search_point_END}{search end point}

\item{cutStart}{cut from start}

\item{cutEnd}{cut from end}

\item{adjust_distal_polyA_end}{adjust distal site or not}

\item{background}{how to get the local background}

\item{z2s}{output of \code{\link[=zScoreThreshold]{zScoreThreshold()}}}

\item{coverage_threshold}{cutoff value for coverage}

\item{long_coverage_threshold}{cutoff value for long form}

\item{PolyA_PWM}{polyA PWM}

\item{classifier}{classifier}

\item{classifier_cutoff}{classifier cutoff}

\item{shift_range}{shift range}

\item{depth.weight}{output of \code{\link[=depthWeight]{depthWeight()}}}

\item{genome}{A \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object}

\item{step}{Adjust step, default 1, means adjust by each base by
cleanUpdTSeq.}

\item{two_way}{Search the proximal site from both direction or not}

\item{tmpfolder}{temp folder could save and reload the analysis data for
resume analysis}

\item{silence}{logical(1), indicating whether progress is reported or not. By
default, FALSE}
}
\value{
a data frame
}
\description{
Estimate the CP sites for UTRs on a giving chromosome
}
\seealso{
\code{\link[=CPsites]{CPsites()}},\code{\link[=searchProximalCPs]{searchProximalCPs()}}, \code{\link[=proximalAdj]{proximalAdj()}},
\code{\link[=proximalAdjByPWM]{proximalAdjByPWM()}}, \code{\link[=proximalAdjByCleanUpdTSeq]{proximalAdjByCleanUpdTSeq()}},\code{\link[=PAscore]{PAscore()}},
\code{\link[=PAscore2]{PAscore2()}}
}
\keyword{internal}
