% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3TotalCoverage.R
\name{fft.smooth}
\alias{fft.smooth}
\title{Smoothing using Fast Discrete Fourier Transform}
\usage{
fft.smooth(sn, p)
}
\arguments{
\item{sn}{}

\item{p}{}
}
\value{

}
\description{
Smoothing using Fast Discrete Fourier Transform
}
\keyword{internal}
