% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherExactTest.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{do fisher exact test for two group datasets}
\usage{
fisherExactTest(UTR3eset, gp1, gp2)
}
\arguments{
\item{UTR3eset}{output of [getUTR3eSet()]}

\item{gp1}{tag names of group 1}

\item{gp2}{tag names of group 2}
}
\value{
a matrix of test results
}
\description{
do fisher exact test for two group datasets
}
\examples{
path <- file.path(find.package("InPAS"), "inst/extdata")
load(file.path(path, "eset.MAQC.rda"))
sample_tags <- colnames(eset$PDUI.log2)    
res <- fisherExactTest(eset, gp1=sample_tags[1:2], 
                       gp2=sample_tags[3:4])

}
\seealso{
[singleSampleAnalyze()],[singleGroupAnalyze()], [limmaAnalyze()]
}
\keyword{internal}
