% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.get_GCandMappability.R
\name{gcComp}
\alias{gcComp}
\title{Calculate weights for GC composition}
\usage{
gcComp(genome, seqnames, window = 50, BPPARAM = NULL)
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{seqnames}{a character(n) vector, the chromosome/scaffolds' names
in the same forms of seqnames in the BSgenome}

\item{window}{size of a sliding window, which optimally is set to the read
length}

\item{BPPARAM}{an optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance
determining the parallel back-end to be used during evaluation, or a list
of BiocParallelParam instances, to be applied in sequence for nested calls
to bplapply. It can be set to NULL or bpparam()}
}
\value{
A list of numeric vectors containing the weight (scaffold-level GC\\%
/ GC\\% per sliding window) for GC composition-based correction for each
chromosome/scaffold.
}
\description{
Calculate read weights for GC composition-based coverage correction
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
genome <- BSgenome.Mmusculus.UCSC.mm10
InPAS:::gcComp(genome, "chr1")
}
}
\references{
Cheung et al. Systematic bias in high-throughput sequencing data
and its correction by BEADS. Nucleic Acids Res. 2011 Aug;39(15):e103.
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
