% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.get_PAscore.R
\name{get_PAscore}
\alias{get_PAscore}
\title{calculate the CP score}
\usage{
get_PAscore(seqname, pos, str, idx, PWM, genome, ups = 50, dws = 50)
}
\arguments{
\item{seqname}{a character(n) vector, the chromosome/scaffold' name}

\item{pos}{genomic positions}

\item{str}{DNA strand}

\item{idx}{offset position}

\item{PWM}{An R object for a position weight matrix (PWM) for a hexamer
polyadenylation signal (PAS), such as AAUAAA.}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{ups}{the number of upstream bases for PAS search.}

\item{dws}{the number of downstream bases for PAS search.}
}
\value{
A list containing offset positions after PA score-based filtering
}
\description{
calculate the CP score by using PWM of polyadenylation signal
}
\seealso{
\code{\link[=get_PAscore2]{get_PAscore2()}}
}
\author{
Jianhong Ou
}
\keyword{internal}
