% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.get_PAscore2.R
\name{get_PAscore2}
\alias{get_PAscore2}
\title{calculate the CP score}
\usage{
get_PAscore2(
  seqname,
  pos,
  str,
  idx,
  idx.gp,
  genome,
  classifier,
  classifier_cutoff,
  future.chunk.size = NULL
)
}
\arguments{
\item{seqname}{a character(1) vector, the chromosome/scaffold's name}

\item{pos}{genomic positions}

\item{str}{DNA strand}

\item{idx}{offset position}

\item{idx.gp}{group number of the offset position}

\item{genome}{an object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{classifier}{An R object for Naive Bayes classifier model, like the one
in the cleanUpdTSeq package.}

\item{classifier_cutoff}{A numeric(1) vector. A cutoff of probability that a
site is classified as true CP sites. The value should be between 0.5 and 1.
Default, 0.8.}

\item{future.chunk.size}{The average number of elements per future
("chunk"). If Inf, then all elements are processed in a single future.
If NULL, then argument future.scheduling = 1 is used by default. Users can
set future.chunk.size = total number of elements/number of cores set for
the backend. See the future.apply package for details.}
}
\value{
a data frame
}
\description{
calculate CP score by cleanUpdTSeq
}
\seealso{
\code{\link[=get_PAscore]{get_PAscore()}}
}
\author{
Jianhong Ou
}
\keyword{internal}
