% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.get_seqLen.R
\name{get_seqLen}
\alias{get_seqLen}
\title{Get sequence lengths for chromosomes/scaffolds}
\usage{
get_seqLen(genome, removeScaffolds = FALSE)
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{removeScaffolds}{A logical(1) vector, whether the scaffolds should be
removed from the genome. If you use a TxDb containing alternative
scaffolds, you'd better to remove the scaffolds. To make things easy, we
suggest users creating a \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} instance from the
reference genome used for read alignment. For details, see the
documentation of \code{\link[BSgenome:BSgenomeForge]{BSgenome::forgeBSgenomeDataPkg()}}.}
}
\value{
A named numeric vector containing lengths per seqname, with the
seqnames as the names
}
\description{
Get sequence lengths for chromosomes/scaffolds from a
\link[BSgenome:BSgenome-class]{BSgenome::BSgenome} object
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
genome <- BSgenome.Mmusculus.UCSC.mm10
InPAS:::get_seqLen(genome, removeScaffolds = FALSE)
}
\seealso{
\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo}
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
