% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getCov}
\title{extract coverage from bedgraph file
 
 extract coverage from bedgraph file}
\usage{
getCov(bedgraph, genome, seqLen, removeScaffolds = FALSE)
}
\arguments{
\item{bedgraph}{bedGraph file names}

\item{genome}{an object [BSgenome::BSgenome-class]}

\item{seqLen}{a named vector containing lengths of each chromosome.
output from [seqLen()]}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome}
}
\value{
an object of [S4Vectors::Rle-class] for a sample coverage
}
\description{
extract coverage from bedgraph file
 
 extract coverage from bedgraph file
}
\examples{

}
\seealso{
[coverageFromBedGraph()]
}
