% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionCoverage.R
\name{getRegionCoverage}
\alias{getRegionCoverage}
\title{calculate coverage for a given region}
\usage{
getRegionCoverage(chr, utr3.regions.chr, hugeData, coverage, phmm = FALSE)
}
\arguments{
\item{chr}{chromosome}

\item{utr3.regions.chr}{the GRanges of region to be extracted}

\item{hugeData}{is it a huge dataset?}

\item{coverage}{output of coverageFromBedGraph}

\item{phmm}{prepare data for singleSample analysis?}
}
\value{
GRanges with coverage data
}
\description{
calculate coverage for a given region
}
\keyword{internal}
