\name{getUTR3region}
\alias{getUTR3region}
\title{
    extract long and short 3UTR region
}
\description{
    extract long and short 3UTR region
}
\usage{
getUTR3region(.grs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.grs}{
    output of CPsites
}
}
\value{
GRanges with short form and long form
}
\author{
    Jianhong Ou
}
\keyword{ misc }