\name{InPAS-package}
\alias{InPAS-package}
\alias{InPAS}
\docType{package}
\title{
    alternative polyadenylation and cleavage estimations
}
\description{
    predict and estimate the alternative polyadenylation and cleavage site
    for mRNA-seq data
}
\details{
    \tabular{ll}{
        Package: \tab InPAS\cr
        Type: \tab Package\cr
        Version: \tab 1.0\cr
        Date: \tab 2014-09-12\cr
        License: \tab GPL (>= 2)\cr
    }
}
\author{
    Jianhong Ou, Sung Mi Park, Michael R. Green and Lihua Julie Zhu

    Maintainer: Jianhong Ou <jianhong.ou@umassmed.edu>
}
\keyword{ package }