% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPAS-package.R
\docType{package}
\name{InPAS}
\alias{InPAS}
\title{A package for the identification of novel Alternative PolyAdenylation
Sites (PAS) based on RNA-seq data}
\description{
The InPAS package provides three categories of important functions:
parseTxDb, utr3Annotation, coverageCDS, coverageFromBedGraph, coverageRate,
getUTR3eSet, utr3UsageEstimation, testUsage, singleSampleAnalysis,
singleGroupAnalysis, limmaAnalysis, filterRes, usage4plot, prepare4GSEA, inPAS
}
\section{functions for retrieving 3' UTR annotation}{
 parseTxDb,
utr3Annotation
}

\section{functions for processing read coverage data}{
 coverageCDS,
coverageFromBedGraph, coverageRate
}

\section{functions for alternative polyadenylation site analysis}{

utr3UsageEstimation, testUsage, singleSampleAnalysis, singleGroupAnalysis,
limmaAnalysis, filterRes, usage4plot
}

