% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.InPAS-package.R
\docType{package}
\name{InPAS}
\alias{InPAS}
\title{A package for identifying novel Alternative PolyAdenylation
Sites (PAS) based on RNA-seq data}
\description{
The InPAS package provides three categories of important functions:
parse_TxDb, extract_UTR3Anno, assemble_allCov, get_ssRleCov, run_coverageQC,
get_UTR3eSet, test_dPDUI, run_singleSampleAnalysis,
run_singleGroupAnalysis, run_limmaAnalysis, filter_testOut,
get_usage4plot, setup_GSEA
}
\section{functions for retrieving 3' UTR annotation}{
 parse_TxDb,
extract_UTR3Anno
}

\section{functions for processing read coverage data}{
 assemble_allCov,
get_ssRleCov, run_coverageQC
}

\section{functions for alternative polyadenylation site analysis}{

test_dPDUI, run_singleSampleAnalysis, run_singleGroupAnalysis,
run_limmaAnalysis, filter_testOut, get_usage4plot
}

