% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covThreshold.R
\name{intronRegion}
\alias{intronRegion}
\title{Helper function to get intronic regions}
\usage{
intronRegion(txdb)
}
\arguments{
\item{txdb}{An object of [GenomicFeatures::TxDb-class]}
}
\value{
GRanges of intronic regions not overlapping exons
}
\description{
Helper function to get intronic regions
}
\keyword{internal}
