% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastCDSusage.R
\name{lastCDSusage}
\alias{lastCDSusage}
\title{extract coverage of last CDS exon region}
\usage{
lastCDSusage(CDS, coverage, hugeData, BPPARAM = NULL, phmm = FALSE)
}
\arguments{
\item{CDS}{a [GenomicRanges::GRanges-class] object for CDS regions}

\item{coverage}{output of [coverageFromBedGraph()]}

\item{hugeData}{logical(1), indicating whether the input is a huge dataset}

\item{BPPARAM}{An optional [BiocParallel::BiocParallelParam-class]
instance determining the parallel back-end to be used during 
evaluation, or a list of BiocParallelParam instances, to be 
applied in sequence for nested calls to bplapply.}

\item{phmm}{logical(1), indicating whether output data is prepared for singleSample analysis}
}
\value{
the average coverage of last CDS for each transcript
}
\description{
extract coverage of last CDS exon region
}
\examples{

}
\keyword{internal}
