% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.get_GCandMappability.R
\name{mapComp}
\alias{mapComp}
\title{Calculate weights for mappability-base coverage correction}
\usage{
mapComp(mi)
}
\arguments{
\item{mi}{A numeric vector of mappability along per
chromosome/scaffold}
}
\value{
A numeric vector of weights for mappability-based
coverage correction
}
\description{
mappability is calculated by using
\href{http://algorithms.cnag.cat/wiki/Man:gem-mappability}{GEM}
with the following command lines:
PATH=$PATH:~/bin/GEM-binaries-Linux-x86_64-core_i3-20130406-045632/bin
./gem-indexer -i genome.fa -o mm10.index.gem
./gem-mappability -I mm10.index.gem.gem -l 100 -o mm10.mappability
./gem-2-wig -I mm10.index.gem.gem -i mm10.mappability -o mm10.mappability.wig
}
\details{
Calculate weights for mappability-base coverage correction
}
\references{
Derrien et al. Fast computation and applications of genome
mappability. PLoS One. 2012;7(1):e30377. doi: 10.1371/journal.pone.0030377.
}
\author{
Jianhong Ou
}
\keyword{internal}
