% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCoverage.R
\name{mergeCoverage}
\alias{mergeCoverage}
\title{Title}
\usage{
mergeCoverage(
  coverage,
  groupList,
  genome,
  removeScaffolds = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{coverage}{coverage for each sample, output of [coverageFromBedGraph()]}

\item{groupList}{group list of tag names}

\item{genome}{an object of [BSgenome::BSgenome-class]}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome}

\item{BPPARAM}{an optional [BiocParallel::BiocParallelParam-class] 
instance determining the parallel back-end to be used during evaluation, 
or a list of BiocParallelParam instances, to be applied in sequence for
nested calls to bplapply}
}
\value{

}
\description{
Title
}
\examples{

}
\keyword{internal}
