% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTxDb.R
\name{parseTxDb}
\alias{parseTxDb}
\title{Extract gene models from a TxDb object}
\usage{
parseTxDb(TxDb = NULL, edb = NULL, removeScaffolds = FALSE)
}
\arguments{
\item{TxDb}{An object of [GenomicFeatures::TxDb-class]}

\item{edb}{An object of [ensembldb::EnsDb-class]}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome. If you use a TxDb containing alternative
scaffolds, you'd better to remove the scaffolds.}
}
\value{
A [GenomicRanges::GRanges-class] object for gene models
}
\description{
Extract gene models from a TxDb object.
}
\details{
A good practice is to perform read alignment using a 
reference genome from Ensembl/GenCode including only the primary
assembly and build a TxDb using the GTF/GFF files downloaded
from the same source as the reference genome, such as 
BioMart/Ensembl/GenCode. For instruction, see Vignette of the 
GenomicFeatures. The UCSC reference genomes and their 
annotation can be very cubersome.
}
\author{
Haibo Liu
}
