% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17.polish_CPs.R
\name{polish_CPs}
\alias{polish_CPs}
\title{polish the searching results of CP sites}
\usage{
polish_CPs(CPs)
}
\arguments{
\item{CPs}{output of \code{\link[=search_proximalCPs]{search_proximalCPs()}} or \code{\link[=adjust_proximalCPs]{adjust_proximalCPs()}}}
}
\value{
a matrix with columns: "fit_value", "Predicted_Proximal_APA",
"Predicted_Distal_APA", "utr3start", "utr3end", "distalCPtype"
}
\description{
remove the multiple positions of CP sites for the same 3' UTRs and only keep
the best CP sites for proximal and distal.
}
\seealso{
\code{\link[=adjust_proximalCPs]{adjust_proximalCPs()}}, \code{\link[=adjust_proximalCPsByPWM]{adjust_proximalCPsByPWM()}},
\code{\link[=adjust_proximalCPsByNBC]{adjust_proximalCPsByNBC()}}, \code{\link[=get_PAscore2]{get_PAscore2()}}
}
\author{
Jianhong Ou
}
\keyword{internal}
