% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polishCPs.R
\name{polishCPs}
\alias{polishCPs}
\title{polish the searching results of CP sites}
\usage{
polishCPs(CPs)
}
\arguments{
\item{CPs}{output of [searchProximalCPs()] or [proximalAdj()]}
}
\value{
a matrix with columns: "fit_value", 
"Predicted_Proximal_APA", "Predicted_Distal_APA", 
"utr3start", "utr3end", "type"
}
\description{
remove the multiple positions of CP sites for same 3UTRs and 
only keep the best CP sites for proximal and distal.
}
\seealso{
[estimateCPsites()], [proximalAdj()], [proximalAdjByPWM()], 
[proximalAdjByCleanUpdTSeq()], [PAscore2()]
}
\keyword{internal}
