% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchProximalCPs.R
\name{searchProximalCPs}
\alias{searchProximalCPs}
\title{search proximal CPsites}
\usage{
searchProximalCPs(
  CPs,
  curr_UTR,
  window_size,
  MINSIZE,
  cutEnd,
  search_point_START,
  search_point_END,
  two_way = FALSE
)
}
\arguments{
\item{CPs}{output from [searchDistalCPs()] or [distalAdj()]}

\item{curr_UTR}{GRanges for current 3' UTR}

\item{window_size}{window size}

\item{MINSIZE}{MINSIZE for short form}

\item{cutEnd}{a numeric(1) between 0 and 1. The percentage of 
nucleotides should be removed from the end before search, 
0.1 means 10 percent.}

\item{search_point_START}{start point for searching}

\item{search_point_END}{end point for searching}

\item{two_way}{Search the proximal site from both direction 
or not.}
}
\value{
a list
}
\description{
search proximal CPsites
}
\examples{
1
}
\seealso{
[proximalAdj()], [polishCPs()],[proximalAdjByPWM()], 
[proximalAdjByCleanUpdTSeq()], [PAscore()], [PAscore2()]
}
\keyword{internal}
