\name{seqLen}
\alias{seqLen}
\title{
get sequence lengths
}
\description{
get sequence lengths from a BSgenome object
}
\usage{
seqLen(genome)
}
\arguments{
  \item{genome}{
    an object of \link[BSgenome:BSgenome-class]{BSgenome}
}
}
\value{
    a numeric vector
}
\author{
    Jianhong Ou
}
\seealso{
    \code{\link[GenomeInfoDb:Seqinfo-class]{seqlengths}}
}
\keyword{ misc }