% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLen.R
\name{seqLen}
\alias{seqLen}
\title{Get sequence lengths}
\usage{
seqLen(genome, removeScaffolds = FALSE)
}
\arguments{
\item{genome}{An object of [BSgenome::BSgenome-class]}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome}
}
\value{
A numeric vector containing lengths per seqnames
}
\description{
Get sequence lengths from a [BSgenome::BSgenome-class] object
}
\examples{
1
}
\seealso{
[GenomeInfoDb::Seqinfo-class]
}
\keyword{internal}
