% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUsage.R
\name{testUsage}
\alias{testUsage}
\title{do test for dPDUI}
\usage{
testUsage(
  CPsites,
  coverage,
  genome,
  utr3,
  BPPARAM = NULL,
  method = c("limma", "fisher.exact", "singleSample", "singleGroup"),
  normalize = c("none", "quantiles", "quantiles.robust", "mean", "median"),
  design,
  contrast.matrix,
  coef = 1,
  robust = FALSE,
  ...,
  gp1,
  gp2
)
}
\arguments{
\item{CPsites}{outputs of [CPsites()]}

\item{coverage}{coverage for each sample, outputs of [coverageFromBedGraph()]}

\item{genome}{an object of [BSgenome::BSgenome-class]}

\item{utr3}{output of [utr3Annotation()]}

\item{BPPARAM}{an optional [BiocParallel::BiocParallelParam-class] 
instance determining the parallel back-end to be used during 
evaluation, or a list of BiocParallelParam instances, to be applied 
in sequence for nested calls to bplapply.}

\item{method}{the test method. see [singleSampleAnalyze()],
[singleGroupAnalyze()],[fisherExactTest()], [limmaAnalyze()]}

\item{normalize}{the normalization method}

\item{design}{a design matrix of the experiment, with rows corresponding 
to arrays and columns to coefficients to be estimated. Defaults to 
the unit vector meaning that the arrays are treated as replicates. 
see [stats::model.matrix()].}

\item{contrast.matrix}{a numeric matrix with rows corresponding to 
coefficients in fit and columns containing contrasts. May be a 
vector if there is only one contrast. see [limma::makeContrasts()]}

\item{coef}{column number or column name specifying which
coefficient or contrast of the linear model is of interest. 
see more [limma::topTable()]. default value: 1}

\item{robust}{logical, should the estimation of the empirical
Bayes prior parameters be robustified against outlier sample
variances?}

\item{...}{other arguments are passed to lmFit}

\item{gp1}{tag names involved in group 1}

\item{gp2}{tag names involved in group 2}
}
\value{
a list with test results in a matrix.
}
\description{
do test for dPDUI
}
\details{
if method is "limma", design matrix and contrast 
is required. if method is "fisher.exact", gp1 and gp2 is required.
}
\examples{
library(limma)
path <- file.path(find.package("InPAS"), "extdata")
load(file.path(path, "CPs.MAQC.rda"))
load(file.path(path, "coverage.MAQC.rda"))
library(BSgenome.Hsapiens.UCSC.hg19)
data(utr3.hg19)
tags <- names(coverage)
g <- factor(gsub("\\\\\\\\..*$", "", tags))
design <- model.matrix(~-1+g)
colnames(design) <- c("Brain", "UHR")
contrast.matrix<-makeContrasts(contrasts="Brain-UHR",levels=design)
res <- testUsage(CPsites=CPs, 
                 coverage=coverage, 
                 genome=BSgenome.Hsapiens.UCSC.hg19,
                 utr3=utr3.hg19, 
                 method="limma",
                 design=design,
                 contrast.matrix=contrast.matrix)
}
\seealso{
[singleSampleAnalyze()], [singleGroupAnalyze()],
[fisherExactTest()], [limmaAnalyze()]
}
