% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalCoverage.R
\name{totalCoverage}
\alias{totalCoverage}
\title{total coverage}
\usage{
totalCoverage(
  coverage,
  genome,
  hugeData = FALSE,
  removeScaffolds = FALSE,
  groupList = NULL
)
}
\arguments{
\item{coverage}{coverage for each sample, outputs of 
[coverageFromBedGraph()]}

\item{genome}{an object of [BSgenome::BSgenome-class]}

\item{hugeData}{hugeData or not}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome}

\item{groupList}{tag names involved in each groups}
}
\value{
a coverage list
}
\description{
for huge dataset, it will read in the coverage from temp files 
and merge them by groups
}
\examples{
1

}
\keyword{internal}
