% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.trim_seqnames.R
\name{trim_seqnames}
\alias{trim_seqnames}
\title{Filter sequence names from a BSgenome object}
\usage{
trim_seqnames(genome = getInPASGenome(), chr2exclude = getChr2Exclude())
}
\arguments{
\item{genome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}}

\item{chr2exclude}{A character vector, NA or NULL, specifying chromosomes or
scaffolds to be excluded for InPAS analysis. \code{chrM} and alternative scaffolds
representing different haplotypes should be excluded.}
}
\value{
An character vector containing filtered seqnames
}
\description{
Filter sequence names for scaffolds from a BSgenome object so that only
chromosome-level seqnames are kept.
}
\author{
Jianhong Ou, Haibo Liu
}
\keyword{internal}
