% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.data.R
\docType{data}
\name{utr3.mm10}
\alias{utr3.mm10}
\title{Annotation of 3' UTRs for mouse (mm10)}
\format{
An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} with 7 metadata columns
\describe{
\item{feature}{feature type, utr3, CDS, next.exon.gap}
\item{annotatedProximalCP}{candidate proximal CPsites}
\item{exon}{exon ID}
\item{transcript}{transcript ID}
\item{gene}{gene ID}
\item{symbol}{gene symbol}
\item{truncated}{whether the 3' UTR is trucated}
}
}
\usage{
utr3.mm10
}
\description{
A dataset containing the annotation of the 3' UTRs of the mouse
}
\keyword{datasets}
