% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr3Annotation.R
\name{utr3Annotation}
\alias{utr3Annotation}
\title{extract 3' UTR information from a [GenomicFeatures::TxDb-class] object}
\usage{
utr3Annotation(
  TxDb = NULL,
  edb = NULL,
  removeScaffolds = FALSE,
  MAX_EXONS_GAP = 10000
)
}
\arguments{
\item{TxDb}{an object of [GenomicFeatures::TxDb-class]}

\item{edb}{An object of [ensembldb::EnsDb-class]}

\item{removeScaffolds}{logical(1), whether the scaffolds should be 
removed from the genome}

\item{MAX_EXONS_GAP}{maximal gap sizes for distal CP site to downstream
exons}
}
\value{
an object of [GenomicRanges::GRanges-class] with 7 metadata columns:
feature (utr3, next.exon.gap, CDS), 
annotatedProximalCP (unknown, proximalCP_<coordinate>), 
exon (<transcript id>_<index>), transcript, 
gene (entrez_id), symbol, truncated (logical).
}
\description{
extract 3' UTR information from a [GenomicFeatures::TxDb-class] 
object. The 3'UTR is defined as the last 3'UTR fragment for 
each transcript and it will be cut if there is any overlaps 
with other exons.
}
\details{
A good practice is to perform read alignment using a 
reference genome from Ensembl/GenCode including only the primary
assembly and build a TxDb using the GTF/GFF files downloaded
from the same source as the reference genome, such as 
BioMart/Ensembl/GenCode. For instruction, see Vignette of the 
GenomicFeatures. The UCSC reference genomes and their 
annotation can be very cubersome.
}
\examples{
if (interactive()){
library(EnsDb.Hsapiens.v86)
TxDb <-loadDb("C:\\\\Users\\\\Haibo\\\\Desktop/003.MAQC.brain.UHR.TxDb.sqlite")
edb <- EnsDb.Hsapiens.v86
tx <- parseEnsDb(TxDb, edb)
utr3Annotation(tx)
}
}
\author{
Jianhong Ou, Haibo Liu
}
