% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3eSet.R
\docType{class}
\name{UTR3eSet-class}
\alias{UTR3eSet-class}
\alias{new}
\alias{UTR3eSet}
\alias{$}
\alias{$,UTR3eSet-method}
\alias{$<-}
\alias{$<-,UTR3eSet-method}
\alias{as}
\alias{show}
\alias{show,UTR3eSet-method}
\title{UTR3eSet-class
 
An object of class [UTR3eSet-class] represents the results of 3' 
UTR usage}
\usage{
UTR3eSet(...)

\S4method{$}{UTR3eSet}(x, name)

\S4method{$}{UTR3eSet}(x, name) <- value

\S4method{show}{UTR3eSet}(object)
}
\arguments{
\item{...}{data for constructing an object of [UTR3eSet-class]}

\item{x}{an object of [UTR3eSet-class]}

\item{name}{slot name}

\item{value}{a value to be assigned}

\item{object}{an object}

\item{UTR3eSet}{an object of [UTR3eSet-class]}

\item{from}{an object of [UTR3eSet-class]}

\item{to}{an object of GRanges}
}
\value{


an object of [UTR3eSet-class]

a slot value of an object of UTR3eSet class
}
\description{
UTR3eSet-class
 
An object of class [UTR3eSet-class] represents the results of 3' 
UTR usage

Constructor

Accessor function for UTR3eSet class

Set slot information of UTR3eSet

convert an object of UTR3eSet class to an ExpressionSet object

convert an object of UTR3eSet class to a GRanges object

print method for UTR3eSet-class
}
\section{Slots}{

\describe{
\item{\code{usage}}{GRanges.}

\item{\code{PDUI}}{matrix.}

\item{\code{PDUI.log2}}{matrix.}

\item{\code{short}}{matrix.}

\item{\code{long}}{matrix.}

\item{\code{signals}}{list.}

\item{\code{testRes}}{matrix.}
}}

