% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR3TotalCoverage.R
\name{UTR3TotalCoverage}
\alias{UTR3TotalCoverage}
\title{extract coverage of 3UTR for CP sites prediction}
\usage{
UTR3TotalCoverage(
  utr3,
  totalCov,
  gcCompensation = NA,
  mappabilityCompensation = NA,
  FFT = FALSE,
  fft.sm.power = 20
)
}
\arguments{
\item{utr3}{an [GenomicRanges::GRanges-class] object. It must be the output of [utr3Annotation()]}

\item{totalCov}{total coverage for each sample. It must be the output of [totalCoverage()]}

\item{gcCompensation}{GC compensation vector. Not support yet.}

\item{mappabilityCompensation}{mappability compensation vector. Not support yet.}

\item{FFT}{Use FFT smooth or not.}

\item{fft.sm.power}{the cut-off frequency of FFT smooth.}
}
\value{
a list. level 1: chromosome; level 2: each transcripts;
level3: data matrix
}
\description{
extract 3UTR coverage from totalCov according to the [GenomicRanges::GRanges-class] object utr3.
}
\keyword{internal}
