% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScoreThreshold.R
\name{zScoreThreshold}
\alias{zScoreThreshold}
\title{Calculate local background cutoff value}
\usage{
zScoreThreshold(background, introns, totalCov, utr3, z = 2)
}
\arguments{
\item{background}{character(1), indicating how background coverage is defined}

\item{introns}{A [GenomicRanges::GRanges-class] object for introns}

\item{totalCov}{total coverage, output from [totalCoverage()]}

\item{utr3}{output of [utr3Annotation()]}

\item{z}{z score cutoff value}
}
\value{
a numeric vector
}
\description{
calculate local background cutoff value based on z-score
}
\examples{
1
}
\keyword{internal}
